% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinysuspend-package.R
\docType{package}
\name{shinysuspend-package}
\alias{shinysuspend}
\alias{shinysuspend-package}
\title{shinysuspend: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
