% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspend.R
\name{suspendInput}
\alias{suspendInput}
\alias{resumeInput}
\alias{suspendForNextFlush}
\title{Suspend or resume changes to a Shiny input}
\usage{
suspendInput(inputId, session = getDefaultReactiveDomain())

resumeInput(inputId, session = getDefaultReactiveDomain())

suspendForNextFlush(inputId, session = getDefaultReactiveDomain())
}
\arguments{
\item{inputId}{The id of the input object.}

\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}
}
\description{
Suspended inputs are prevented from changing their values visible to Shiny.
Suspending and resuming take effect immediately. \code{suspendForNextFlush()} is
a convenience wrapper to suspend an input for the duration of a single flush
cycle, useful for e.g. avoiding invalidations by \verb{update*()} functions.
}
