% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinysuspend-package.R
\name{useShinysuspend}
\alias{useShinysuspend}
\title{Include dependencies for shinysuspend in a Shiny app}
\usage{
useShinysuspend()
}
\description{
Include dependencies for shinysuspend in a Shiny app
}
