% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinysuspend-package.R
\docType{package}
\name{shinysuspend-package}
\alias{shinysuspend}
\alias{shinysuspend-package}
\title{shinysuspend: Suspend Shiny Inputs}
\description{
Suspend (and resume) Shiny inputs so that changes to their values are not propagated to the Shiny session.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikmart/shinysuspend}
  \item Report bugs at \url{https://github.com/mikmart/shinysuspend/issues}
}

}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
